+++ dataset/HR_5/HR-5_v012/fast.c
@@ -4,6 +4,6 @@
 #include <string.h>
 
 __attribute__((noinline))
-void fast_hr5_v012(int *out, int *A, int *B, int n) {
+void fast_hr5_v012(double *out, double *A, double *B, int n) {
     for (int i = 0; i < n; i++) out[i] = A[i] + B[i];
 }
