+++ dataset/MI_4/MI-4_v015/fast.c
@@ -4,7 +4,7 @@
 #include <string.h>
 
 __attribute__((noinline))
-void fast_mi4_v015(int *out, int *A, int *B, int rows, int cols) {
+void fast_mi4_v015(double *out, double *A, double *B, int rows, int cols) {
     for (int i = 0; i < rows; i++) {
         for (int j = 0; j < cols; j++) {
             out[i * cols + j] = A[i * cols + j] + B[i * cols + j];
