+++ dataset/MI_4/MI-4_v015/slow.c
@@ -4,7 +4,7 @@
 #include <string.h>
 
 __attribute__((noinline))
-void slow_mi4_v015(int *out, int *A, int *B, int rows, int cols) {
+void slow_mi4_v015(double *out, double *A, double *B, int rows, int cols) {
     for (int j = 0; j < cols; j++) {
         for (int i = 0; i < rows; i++) {
             out[i * cols + j] = A[i * cols + j] + B[i * cols + j];
